% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_correction.R
\name{blankErr}
\alias{blankErr}
\title{Determine error in blankfm}
\usage{
blankErr(blankfm, blankerr = NULL)
}
\arguments{
\item{blankfm}{A vector of normalized Fm's of blanks.}

\item{blankerr}{An optional vector of normalized Fm errors of blanks.}
}
\value{
An error for the given blanks.
}
\description{
Uses the "error floor" method from SNICSer
}
