% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reduction.R
\name{calc_d13c}
\alias{calc_d13c}
\title{Calculate d13C}
\usage{
calc_d13c(r_unk, r_vpdb = 0.0112372)
}
\arguments{
\item{r_unk}{13/12 C ratio of the unknowns}

\item{r_vpdb}{13/12 C ratio of VPDB}
}
\value{
A vector of d13C values
}
\description{
Calculate d13C
}
