% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blank_correction.R
\name{doLBC}
\alias{doLBC}
\title{Apply large blank}
\usage{
doLBC(fmmeas, fmblank, fmstd)
}
\arguments{
\item{fmmeas}{Normalized Fm of the sample}

\item{fmblank}{Normalized Fm of the blank}

\item{fmstd}{Accepted Fm of the standard used for normalization}
}
\value{
Blank corrected Fm
}
\description{
Apply large blank
}
