% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_standards.R
\name{find_standard_runs}
\alias{find_standard_runs}
\title{Create an index of standard runs}
\usage{
find_standard_runs(types, drop = NULL)
}
\arguments{
\item{types}{A character vector of sample type ids where "S" denotes a normalizing standard}

\item{drop}{The index of a runs to ignore (to avoid self-normalization)}
}
\value{
The indexes of standard runs
}
\description{
Create an index of standard runs
}
