% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_standards.R
\name{nearest_standards}
\alias{nearest_standards}
\title{Find nearest standards}
\usage{
nearest_standards(run, stds, n)
}
\arguments{
\item{run}{run index}

\item{stds}{indexes of standard runs}

\item{n}{number of standard runs}
}
\value{
A vector of nearest standard run indexes
}
\description{
Find nearest standards
}
