% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_standards.R
\name{nearest_standards_time}
\alias{nearest_standards_time}
\title{Find nearest standards in time}
\usage{
nearest_standards_time(types, times, run, n, self = FALSE)
}
\arguments{
\item{types}{A character vector of sample type ids where "S" denotes a normalizing standard}

\item{times}{A vector of run timestamps}

\item{run}{The index of the run to find standards for}

\item{n}{number of standard runs}

\item{self}{Allow self normalization}
}
\value{

}
\description{
Find nearest standards in time
}
