% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_snicser}
\alias{norm_snicser}
\title{Calculate normalized ratio of a run, SNICSer style}
\usage{
norm_snicser(run_val, std_vals, std_rat = 1.0398)
}
\arguments{
\item{run_val}{The ratio of the run to normalize}

\item{std_vals}{A vector of standard ratios to normalize with}

\item{std_rat}{The standard ratio(s). May be one value for all stds
used or a vector of values for each standard in std_vals, defaults
to OX-I}
}
\value{
The normalized ratio of the run
}
\description{
Snicser does things in this order:
}
\details{
1. 13C correction (prior to this fn)
2. Divide ratio of stds by known standard Fm
3. Average standard ratios
4. Divide run ratio by normalized avg std ratio
}
