% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_snicser_err}
\alias{norm_snicser_err}
\title{Calculate error of a run, SNICSer style}
\usage{
norm_snicser_err(run_val, run_err, std_vals, std_errs, std_rat = 1.0398)
}
\arguments{
\item{run_val}{The ratio of the run to normalize}

\item{run_err}{The error of the run to normalize}

\item{std_vals}{A vector of standard ratios to normalize with}

\item{std_errs}{A vector of errors of the standard runs}

\item{std_rat}{The standard ratio(s). May be one value for all stds
used or a vector of values for each standard in std_vals, defaults
to OX-I}
}
\value{
The error in the normalized ratio of the run
}
\description{
Calculate error of a run, SNICSer style
}
