% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{rc_age_err}
\alias{rc_age_err}
\title{Calculate radiocarbon age error}
\usage{
rc_age_err(fm, fm_err)
}
\arguments{
\item{fm}{Fraction modern}

\item{fm_err}{Error in the Fm}
}
\value{
radiocarbon age error in years
}
\description{
Calculate radiocarbon age error
}
